requirements.txt
git+https://github.com/behave/behave@v1.2.7.dev6#egg=behave
selenium>=4.23.1
webdriver-manager>=4.0.2
geckodriver-autoinstaller>=0.1.0
junitparser>=3.1.2
#authlib>=1.3.1
requests>=2.32.3
beautifulsoup4>=4.12.3
stac-api-validator>=0.6.3
rs-client-libraries==0.2a5

